// frontend/src/api/models/instanceQueryResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { InstanceQueryResultGroupValues } from './instanceQueryResultGroupValues';
import type { InstanceResponse } from './instanceResponse';

export interface InstanceQueryResult {
  groupValues: InstanceQueryResultGroupValues;
  instance: InstanceResponse;
}