// frontend/src/api/models/workflowResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { Schema } from './schema';
import type { WorkflowStatus } from './workflowStatus';
import type { WorkflowUseCase } from './workflowUseCase';

export interface WorkflowResponse {
  createdAt: string;
  instanceCount: number;
  jobId: string | null;
  labelGroupIds: string[];
  novelIds: string[];
  schema: Schema;
  updatedAt: string;
  useCase: WorkflowUseCase;
  workflowId: string;
  workflowMessage: string | null;
  workflowName: string | null;
  workflowStatus: WorkflowStatus;
}