// frontend/src/api/models/workflowSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { Schema } from './schema';
import type { WorkflowStatus } from './workflowStatus';
import type { WorkflowUseCase } from './workflowUseCase';

export interface WorkflowSummary {
  createdAt: string;
  jobId: string | null;
  schema: Schema;
  updatedAt: string;
  useCase: WorkflowUseCase;
  workflowId: string;
  workflowMessage: string | null;
  workflowName: string | null;
  workflowStatus: WorkflowStatus;
}